Convert app to Flutter package
