+++ lib/elastic_list_view.dart
@@ -1,0 +1,3 @@
+library flutter_elastic_list_view;
+
+export 'src/ui/elastic_list_view.dart';
